refactors project to add unit tests
