+++ .travis.yml
@@ -1,0 +1,14 @@
+dist: trusty
+sudo: false
+language: python
+python:
+  - "3.2"
+  - "3.3"
+  - "3.4"
+  - "3.5"
+  - "3.6"
+  - "nightly"
+
+script:
+  - python -m unittest
+
