+++ tests/context.py
@@ -1,0 +1,11 @@
+
+#
+# Inserts module at beginning of sys.path in order to test source code, not
+# installed package.
+#
+
+import os
+import sys
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+import csvsorter
